pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "social-scraper"
version = "0.2.0"
description = "Video extraction, download, and analysis pipeline for TikTok & Instagram"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "playwright>=1.41.0",
    "yt-dlp>=2024.1.0",
    "google-generativeai>=0.4.0",
    "supabase>=2.3.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
